refactor(#205): Param으로 토큰 받도록 변경
